Extract screenshot-checking to the crate root
