+++ uefi-test-runner/src/main.rs
@@ -10,6 +10,8 @@ extern crate log;
 extern crate alloc;
 
 use uefi::prelude::*;
+use uefi::proto::console::serial::Serial;
+use uefi_exts::BootServicesExt;
 
 mod boot;
 mod proto;
@@ -48,6 +50,35 @@ fn check_revision(rev: uefi::table::Revision) {
         minor >= 30,
         "Old version of UEFI 2, some features might not be available."
     );
+}
+
+/// Ask the test runner to check the current screen output against a reference
+/// TODO: This is obviously highly QEMU-specific
+/// TODO: Turn it into something which waits a bit for user inspection elsewhere
+fn check_screenshot(bt: &BootServices, name: &str) {
+    // Access the serial port (in a QEMU environment, it should always be there)
+    let mut serial = bt.find_protocol::<Serial>()
+                       .expect("Could not find serial port");
+    let serial = unsafe { serial.as_mut() };
+
+    // Set a large-ish timeout to avoid problems
+    let mut io_mode = serial.io_mode().clone();
+    io_mode.timeout = 100_000;
+    serial.set_attributes(&io_mode)
+          .expect("Failed to configure serial port");
+
+    // Send a screenshot request to QEMU
+    // TODO: Do not hardcode the screenshot name
+    // TODO: Implement write!() for serial ports
+    let screenshot_request = b"SCREENSHOT: gop_test\n";
+    let write_size = serial.write(screenshot_request).expect("Failed to write screenshot command");
+    assert_eq!(write_size, screenshot_request.len(), "Screenshot request timed out");
+
+    // Wait for QEMU's acknowledgement before moving forward
+    let mut reply = [0; 3];
+    let read_size = serial.read(&mut reply[..]).expect("Failed to read host reply");
+    assert_eq!(read_size, 3, "Screenshot request timed out");
+    assert_eq!(&reply[..], b"OK\n", "Unexpected screenshot request reply");
 }
 
 fn shutdown(st: &SystemTable) -> ! {
